requirements.txt
deeponto
ontospy
openai
owlready2
python-dotenv
rank_bm25
rdflib
